fixes #27; Clean-up display logic and bug correction
Cleaned-up the display logic by reusing an existing variable instead of creating a new one.
Tweaked the logic to correct a behavior where the user could
Open the create fournisseur modal
Enable the edit mode for "numéro"
Close the create fournisseur modal
Open the update fournisseur modal and change the value for "numéro"

Found a bug concerning the focus of the "numéro" element.
If a "numéro" is already used, (which is the case when updating), the focus would be directed toward the input.
That was happening when updating the Fournisseur and thus, messed up the navigation by constantly giving back the focus to the "numéro" input.