fixes #21; Columns now resize like expected.
Previous behavior -> resizing the form wouldn't resize the column width
Current behavior -> resizing the form automatically adjust the column width
Added a property "priority : number" to headerName. Going up in value decrease the priority. (0 being the "highest" priority).
When the component is mounted, it start to listen to window resize.
When it resizes, a visibilityLevel is set by comparing the width of the window to an array of thresholds.
The visibility level is then compared to the priority property of the headerName to exclude some of the columns.
Adding thresholds lets fine tune the column visibility.
Doesn't interfere with the hide property used to switch columns when selecting checkboxes.