fixes #39, fixes #40, fixes #41, fixes #42, fixes #43, fixes #44; New view ResultatsComptables with new functionnalities.
v-select input display a list coming from the backend with the name of the folders (rn, the backend returns a mock-list, though)
First set of DatePicker displayed and is initialized with the starting and ending date of the current Periode Comptable that it fetches through the API.
Second set of DatePicker added. By default, it is hidden and is revealed on the press of a button. Those dates are initialized based on the previous Periode Comptable.
2 textboxes added to allow the user to pick either Bilan, Compte de résultat or both.
Button Chargement added. It is greyed out when you don't have the necessary requirement to create a correct post request to the backend. Clicking it format and send a DTO to the backend. The backend response isn't implemented yet